modified:   RFM12B.cpp 	modified:   RFM12B.h 	modified:   keywords.txt
Added a ClearToSend function to provide information back to the main
program as to the current state of the library.  If the interrupt routine
is busy receiving or sending data, this function returns a false.
This allows the main program to go on doing other tasks instead of
calling the send routine which in turn would just wait until the RF12 is
ready.